api/requirements.txt
numpy
pillow
fastapi
uvicorn
tensorflow
selenium==4.18.1